// src/locations/interfaces/locations.interface.ts
// Generated by https://quicktype.io
export interface DepartmentsAPI {
  id: number;
  name: string;
  description: string;
  cityCapitalId: number;
  municipalities: number;
  surface: number;
  population: number;
  phonePrefix: string;
  countryId: number;
  cityCapital: null;
  country: null;
  cities: null;
  regionId: number;
  region: null;
  naturalAreas: null;
  maps: null;
}

// Generated by https://quicktype.io
export interface CitiesAPI {
  id: number;
  name: string;
  description: string;
  surface: number | null;
  population: number | null;
  postalCode: null | string;
  departmentId: number;
  department: null;
  touristAttractions: null;
  presidents: null;
}

export interface IDepartment {
  id: number;
  name: string;
}

export interface ICity {
  id: number;
  name: string;
  state: IDepartment;
}